demo_facebook_integration.py
"""
Demo Script - Facebook Collection with Best Pages Integration
Tests the integrated Facebook collector using validated best pages
"""

import asyncio
import sys
import os
from datetime import datetime
from dotenv import load_dotenv

# Add project root to path
sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '.')))

from app.collectors.data_pipeline import DataCollectionPipeline
from app.database.connection import get_db
from app.models.tourist_attraction import TouristAttraction
from app.models.province import Province
from app.core.config import settings

# Load environment variables
load_dotenv()

async def test_facebook_integration():
    """Test Facebook collection with best pages through the pipeline"""
    
    print("="*80)
    print("🧪 DEMO: Facebook Collection with Best Pages Integration")
    print("="*80)
    print(f"\n⏰ Started: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    
    # Check configuration
    print("\n📋 Configuration Check:")
    print(f"   Facebook enabled: {settings.FACEBOOK_COLLECTION_ENABLED}")
    print(f"   Use best pages: {settings.FACEBOOK_USE_BEST_PAGES}")
    print(f"   Posts per location: {settings.FACEBOOK_POSTS_PER_LOCATION}")
    print(f"   Comments per post: {settings.FACEBOOK_COMMENTS_PER_POST}")
    
    print("\n🎯 Best Pages Available:")
    for key, config in settings.FACEBOOK_BEST_PAGES.items():
        print(f"   {key}: {config['name']} (~{config['expected_comments_per_post']:.1f} comments/post)")
    
    # Get API token
    apify_token = os.getenv('APIFY_API_TOKEN')
    if not apify_token:
        print("\n❌ APIFY_API_TOKEN not found in environment")
        return
    
    print("\n✅ Apify token found")
    
    # Initialize pipeline
    print("\n🔧 Initializing Data Collection Pipeline...")
    pipeline = DataCollectionPipeline(apify_api_token=apify_token)
    
    if 'facebook' not in pipeline.collectors:
        print("❌ Facebook collector not initialized")
        return
    
    print("✅ Pipeline initialized with Facebook collector")
    
    # Get database session
    db = next(get_db())
    
    # Find test attractions
    print("\n🔍 Finding test attractions...")
    
    test_cases = [
        {"name": "Bà Nà Hills", "province": "Đà Nẵng"},
        {"name": "Đà Lạt", "province": "Lâm Đồng"},
        {"name": "Phú Quốc", "province": "Kiên Giang"}
    ]
    
    for test_case in test_cases:
        attraction_name = test_case["name"]
        province_name = test_case["province"]
        
        print(f"\n{'='*80}")
        print(f"📍 Testing: {attraction_name} ({province_name})")
        print(f"{'='*80}")
        
        # Find or create province
        province = db.query(Province).filter(
            Province.name.ilike(f"%{province_name}%")
        ).first()
        
        if not province:
            print(f"⚠️  Province '{province_name}' not found in database, creating...")
            province = Province(
                name=province_name,
                code=province_name[:2].upper()
            )
            db.add(province)
            db.commit()
            db.refresh(province)
            print(f"✅ Created province: {province.name}")
        
        # Find or create attraction
        attraction = db.query(TouristAttraction).filter(
            TouristAttraction.name.ilike(f"%{attraction_name}%"),
            TouristAttraction.province_id == province.id
        ).first()
        
        if not attraction:
            print(f"⚠️  Attraction '{attraction_name}' not found, creating...")
            attraction = TouristAttraction(
                name=attraction_name,
                province_id=province.id,
                description=f"Demo attraction for testing: {attraction_name}"
            )
            db.add(attraction)
            db.commit()
            db.refresh(attraction)
            print(f"✅ Created attraction: {attraction.name} (ID: {attraction.id})")
        else:
            print(f"✅ Found attraction: {attraction.name} (ID: {attraction.id})")
        
        # Collect data
        print(f"\n🚀 Starting collection for {attraction.name}...")
        print(f"   Limit: {settings.FACEBOOK_POSTS_PER_LOCATION} posts")
        print(f"   Comments: {settings.FACEBOOK_COMMENTS_PER_POST} per post")
        
        try:
            result = await pipeline.collect_for_attraction(
                attraction_id=attraction.id,
                platforms=['facebook'],
                limit_per_platform=settings.FACEBOOK_POSTS_PER_LOCATION
            )
            
            # Display results
            print(f"\n📊 Collection Results:")
            print(f"   Total posts: {result['total_posts']}")
            print(f"   Total comments: {result['total_comments']}")
            
            if 'facebook' in result['platforms']:
                fb_result = result['platforms']['facebook']
                print(f"\n   Facebook Details:")
                print(f"      Strategy: {fb_result.get('strategy', 'N/A')}")
                print(f"      Best page: {fb_result.get('best_page_used', 'N/A')}")
                print(f"      Posts collected: {fb_result['posts_collected']}")
                print(f"      Comments collected: {fb_result['comments_collected']}")
                print(f"      Avg comments/post: {fb_result.get('avg_comments_per_post', 'N/A')}")
                print(f"      Filter efficiency: {fb_result.get('filter_efficiency', 'N/A')}")
            
            if result.get('errors'):
                print(f"\n⚠️  Errors encountered:")
                for error in result['errors']:
                    print(f"      {error}")
            
            print(f"\n✅ Collection completed successfully!")
            
        except Exception as e:
            print(f"\n❌ Collection failed: {str(e)}")
            import traceback
            traceback.print_exc()
        
        # Wait between attractions to avoid rate limiting
        print(f"\n⏸️  Waiting 5 seconds before next attraction...")
        await asyncio.sleep(5)
    
    db.close()
    
    print("\n" + "="*80)
    print("🎉 DEMO COMPLETED!")
    print("="*80)
    print(f"⏰ Finished: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
    print("\n📝 Summary:")
    print("   ✅ Facebook collector integrated with best pages")
    print("   ✅ 2-actor strategy (posts + comments) working")
    print("   ✅ Smart Page Selection validated")
    print("   ✅ Ready for production use!")
    print("\n🚀 Next steps:")
    print("   1. Enable scheduler for automated collection")
    print("   2. Add more attractions to database")
    print("   3. Monitor collection logs and adjust limits")

if __name__ == "__main__":
    asyncio.run(test_facebook_integration())
