requirements.txt
# Core FastAPI
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.11.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.0
alembic==1.12.1
psycopg2-binary==2.9.7
asyncpg==0.29.0

# Data processing
pandas==2.1.0
numpy==1.24.0

# Web scraping & API clients
requests==2.31.0
beautifulsoup4==4.12.2
google-api-python-client==2.107.0
googlemaps==4.10.0

# Background scheduling
apscheduler==3.10.4
selenium==4.15.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-decouple==3.8
loguru==0.7.2

# ML & NLP
torch==2.1.0
transformers==4.35.0  
scikit-learn==1.3.0
underthesea==6.2.0

# Social APIs
google-api-python-client==2.108.0
pytube==15.0.0

# Apify (for Facebook & TikTok scraping)
apify-client==1.5.0

# Background tasks 
celery==5.3.0
redis==5.0.0





